Fix AC_CHECK_HEADERS usage for autoconf 2.69
The fourth argument of AC_CHECK_HEADERS (used to indicate the set of
includes), is mandatory for autoconf 2.69 and earlier to force modern
behavior of the macro.

The semantics of the macro changed from using a preprocessor-only check to
a compiler check, and the fourth argument determines the use of: classic
behavior (for compatibility), a transitional check of both and warnings
(returning retro-compatible but potentially inconsistent results), or the
use of the modern compiler-based test only. Specifying this argument also
suppresses a warning about the topic.

Fixes: 4ada39f9eb61 ("build: fix tirpc usage #27")
Ref: https://github.com/openSUSE/xinetd/pull/40
Signed-off-by: Ismael Luceno <ismael@iodev.co.uk>